+++ themes/learning-management-system/learning-management-theme/src/pages/contactUs.js
@@ -1,6 +1,6 @@
 import React from "react"
-import ContactUs from "../component/contactUs"
 import Layout from "../ui/layout"
+import ContactUs from "../component/contactUs"
 
 const ContactUsPage = () => {
   return (
