+++ themes/learning-management-system/learning-management-theme/src/pages/user.js
@@ -1,0 +1,13 @@
+import React from "react"
+import UserDetail from "../component/userDetail"
+import Layout from "../ui/layout"
+
+const User = () => {
+  return (
+    <Layout>
+      <UserDetail />
+    </Layout>
+  )
+}
+
+export default User
